% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keysight_data_handling.R
\name{adjust_raw_colnames}
\alias{adjust_raw_colnames}
\title{Adjust columns form Keysight data}
\usage{
adjust_raw_colnames(raw_section_list)
}
\arguments{
\item{raw_section_list}{A list of dataframes with each section}
}
\value{
\code{dataframe}
}
\description{
Iterates through all sections
in the raw data file
and names their columns accordingly
}
