% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphotnics_functions.R
\name{cut_back_method}
\alias{cut_back_method}
\title{Cutback Method}
\usage{
cut_back_method(wg_loss_df, grouping)
}
\arguments{
\item{wg_loss_df}{This requires a dataframe with \code{lot}, \code{wafer}, \code{lambda}, \code{loss1}, \code{x_y_die}, & \code{dims}}
}
\value{
\code{dataframe}
}
\description{
Fit Waveguide Insertion Loss
}
