% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keysight_data_handling.R
\name{generate_section_col_names}
\alias{generate_section_col_names}
\title{Generate Section Names}
\usage{
generate_section_col_names(df_section)
}
\arguments{
\item{df_section}{A dataframe containing sections of data}
}
\value{
\code{list}
}
\description{
generate names for each column of data
For example, the TE/TM section is named with the amount
of fibers observed and modes
}
