% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keysight_data_handling.R
\name{generate_te_tm_col_names}
\alias{generate_te_tm_col_names}
\title{Generate Section Names}
\usage{
generate_te_tm_col_names(df_te_tm)
}
\arguments{
\item{df_te_tm}{A dataframe based only on tm_te pattern}
}
\value{
\code{List}
}
\description{
generate names for the TE/TM column of data
For example, the TE/TM section is named with the amount
of fibers observed and modes
}
