% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphotnics_functions.R
\name{get_number_from_string}
\alias{get_number_from_string}
\title{Get Number from string}
\usage{
get_number_from_string(input_string)
}
\arguments{
\item{input_string}{Waveguide dimensions as a string}
}
\value{
\code{int} or \code{double}
}
\description{
Converts a string into an int/double
by decoding the string
}
