% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keysight_data_handling.R
\name{get_te_data}
\alias{get_te_data}
\title{Get TE mode data}
\usage{
get_te_data(df_te_tm, fiber_nums = c())
}
\arguments{
\item{df_te_tm}{dataframe containing te_tm data}

\item{fiber_nums}{Number of optical fibers}
}
\value{
\code{dataframe}
}
\description{
Get TE mode data from
a dataframe that contains te_tm data.
This function allows you to put in specific fibers
for analysis.
}
