% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphotnics_functions.R
\name{lm_metrics}
\alias{lm_metrics}
\title{Fitting Linear Models for perfomance}
\usage{
lm_metrics(data, y_term, x_term)
}
\arguments{
\item{data}{\verb{data table} expected}

\item{y_term}{This varibale will be on LHS of equation in the form:
y = mx + b}

\item{x_term}{This varibale will be on RHS of equation in the form:
y = mx + b}
}
\value{
\verb{data table}
}
\description{
Linear fitting with term coefficients and r-squared
}
