% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphotnics_functions.R
\name{map_files_to_dfs}
\alias{map_files_to_dfs}
\title{Map files to dfs}
\usage{
map_files_to_dfs(directory)
}
\arguments{
\item{directory}{Files that have been collected from Keysight Software}
}
\value{
\code{dataframe}
}
\description{
Creates a \code{FileID} column that contains
the file path and makes a new dataframe
for each file with the new column
}
