% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keysight_data_handling.R
\name{write_sections_to_files}
\alias{write_sections_to_files}
\title{Write Keysight Sections}
\usage{
write_sections_to_files(section_bodys, path)
}
\arguments{
\item{section_bodys}{list sections from keysight data}

\item{path}{Directory for files to be saved}
}
\value{
void
}
\description{
Creates files for each
section contained in the \verb{section body} param
}
