% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keysight_data_handling.R
\name{split_keysight_csv}
\alias{split_keysight_csv}
\title{Keysight Data Filter}
\usage{
split_keysight_csv(raw_data_path, section_delim = "=")
}
\arguments{
\item{raw_data_path}{Data path containg data straight from Keysight tool}

\item{section_delim}{How Keysight data is divided within file}
}
\value{
\code{dataframe}
}
\description{
Split data from Keysight data drop
Keysight tool creates csv files that contain:
fab info
Overview
mueller_row_1_TLS0
average_il_TLS0_
pdl_TLS0_
te_tm_TLS0
}
